Conflicts:
	build.sh

+++ build.sh
@@ -4,7 +4,7 @@ set -o errexit
 TARBALL=$1
 DOCKER=${DOCKER:-docker}
 BASEDIR=$(dirname $(readlink -f ${BASH_SOURCE[0]}))
-NODES=("frontend" "node" "server" "node-colmet" "server-colmet")
+NODES=("frontend" "node" "server")
 TMPDIR=$(mktemp -d --tmpdir docker-oarcluster.XXXXXXXX)
 
 
@@ -16,19 +16,29 @@ fail() {
 
 $DOCKER 2> /dev/null || fail "error: Docker ($DOCKER) executable no found. Make sure Docker is installed and/or use the DOCKER variable to set Docker executable."
 
-. $BASEDIR/clean.sh
-
 [ -n "$TARBALL" ] || fail "error: You must provide a URL to a OAR tarball"
-[ -r "$TARBALL" ] && TARBALL=file://$TARBALL
+[ -r "$TARBALL" ] && TARBALL=file://$(readlink -m $TARBALL)
 
 curl $TARBALL -o $TMPDIR/oar-tarball.tar.gz || fail "error: failed to fetch oar tarball at $TARBALL"
 VERSION=$(tar xfz $TMPDIR/oar-tarball.tar.gz --wildcards "*/sources/core/common-libs/lib/OAR/Version.pm" --to-command "grep -e 'my \$OARVersion'" | sed -e 's/^[^"]\+"\(.\+\)";$/\1/')
 [ -n "${VERSION}" ] || fail "error: fail to retrieve OAR version"
 
+. $BASEDIR/clean.sh
+
+
 # forward OAR version if necessary
 for image in "${NODES[@]}"; do
-    echo "$VERSION" > $BASEDIR/images/$image/version.txt
-    cp $TMPDIR/oar-tarball.tar.gz $BASEDIR/images/$image/
+    if [ -f $BASEDIR/images/$image/oar-tarball.tar.gz ]; then
+        NEW_MD5=$(md5sum $TMPDIR/oar-tarball.tar.gz | awk '{print $1}')
+        OLD_MD5=$(md5sum $BASEDIR/images/$image/oar-tarball.tar.gz | awk '{print $1}')
+        if [ ! "$NEW_MD5" == "$OLD_MD5" ]; then
+            echo "$VERSION" > $BASEDIR/images/$image/version.txt
+            cp $TMPDIR/oar-tarball.tar.gz $BASEDIR/images/$image/
+        fi
+    else
+        echo "$VERSION" > $BASEDIR/images/$image/version.txt
+        cp $TMPDIR/oar-tarball.tar.gz $BASEDIR/images/$image/
+    fi
 done
 
 rm -rf $TMPDIR
